Add tests sent6 & 7. Found bug.
The bug I found is with the Quantifiers. In the Eidos shell, it will attach the Quanitifer as Some(ArraySeq()), instead of Quant(). For example:
Attachments: Increase(increase,Some(ArraySeq(small))) instead of Quant("small").
The grammar does still extract the Quantifier on its own, e.g.:
List(Quantifier) => small