Added a method to deduplicate, but not in use
I left it in bc we have a ton of duplicate events, so we're
expanding unneeded things.  I didn't know if it would be
faster to prune them before expansion, or if the groupBy
itself made it too expensive.  so i left it in in case in
the future it was useful.  we can remove if desired.